requirements.txt
pyside2
pillow
pytest
pytest-cov
